+++ functions/theme-functions.php
@@ -4,7 +4,7 @@
  *
  */
 
-if ( ! function_exists( 'mds_setup' ) ) :
+if (!function_exists( 'mds_setup')) {
     /**
      * Sets up theme defaults and registers support for various WordPress features.
      *
@@ -12,6 +12,7 @@ if ( ! function_exists( 'mds_setup' ) ) :
      * runs before the init hook. The init hook is too late for some features, such
      * as indicating support for post thumbnails.
      */
+    add_action('after_setup_theme', 'mds_setup');
     function mds_setup() {
 
         /**
@@ -20,7 +21,7 @@ if ( ! function_exists( 'mds_setup' ) ) :
          * If you're building a theme based on , use a find and replace
          * to change '' to the name of your theme in all the template files
          */
-        load_theme_textdomain( '', get_template_directory() . '/languages' );
+        load_theme_textdomain('mds_starter_theme', get_template_directory() . '/languages');
 
         /**
          * Let WordPress manage the document title.
@@ -28,60 +29,37 @@ if ( ! function_exists( 'mds_setup' ) ) :
          * hard-coded <title> tag in the document head, and expect WordPress to
          * provide it for us.
          */
-        add_theme_support( 'title-tag' );
+        add_theme_support('title-tag');
 
         /**
          * Enable support for Post Thumbnails on posts and pages.
          *
          */
-        add_theme_support( 'post-thumbnails' );
+        add_theme_support('post-thumbnails');
 
 
         /**
          * Register menu locations
          */
-        register_nav_menus( array(
-            'primary' => __( 'Primary Menu', '' ),
-        ) );
+        register_nav_menus(array(
+            'primary' => __('Primary Menu', ''),
+            'footer' => __('Footer Menu', ''),
+        ));
 
         /**
          * Switch default core markup for search form, comment form, and comments
          * to output valid HTML5.
          */
-        add_theme_support( 'html5', array(
+        add_theme_support('html5', array(
             'search-form',
             'comment-form',
             'comment-list',
             'gallery',
             'caption',
-        ) );
+        ));
 
-        /**
-         * Enable support for Post Formats.
-         */
-        add_theme_support( 'post-formats', array(
-            'aside',
-            'image',
-            'video',
-            'quote',
-            'link',
-        ) );
-
-        /**
-         * Set up the WordPress core custom background feature.
-         */
-        add_theme_support( 'custom-background', apply_filters( '_s_custom_background_args', array(
-            'default-color' => 'ffffff',
-            'default-image' => '',
-        ) ) );
     }
-endif;
-add_action( 'after_setup_theme', 'mds_setup' );
-
-/**
- * Allow wp-editor in Option Tree Metaboxes
- */
-add_filter( 'ot_override_forced_textarea_simple','__return_true' );
+}
 
 /**
  * Set the content width in pixels, based on the theme's design and stylesheet.
@@ -90,11 +68,10 @@ add_filter( 'ot_override_forced_textarea_simple','__return_true' );
  *
  * @global int $content_width
  */
+add_action( 'after_setup_theme', 'mds_content_width', 0 );
 function mds_content_width() {
     $GLOBALS['content_width'] = apply_filters( 'mds_content_width', 640 );
 }
-add_action( 'after_setup_theme', 'mds_content_width', 0 );
-
 
 /**
  * Hide default post/pages admin screen options
@@ -143,20 +120,67 @@ function mds_mime_types($mimes) {
  * 
  * @return array
  */
+add_filter('body_class', 'mds_body_classes');
 function mds_body_classes($classes) {
     if ( ! is_singular() ) {
         $classes[] = '';
     }
     return $classes;
 }
-add_filter( 'body_class', 'mds_body_classes' );
 
 /**
- * Remove width and height attribute from post thumbnail
+ * Get icons file url
+ *
+ * @return string
  */
-add_filter( 'post_thumbnail_html', 'mds_remove_thumbnail_dimensions', 10, 3 );
-function mds_remove_thumbnail_dimensions( $html, $post_id, $post_image_id ) {
-    $html = preg_replace( '/(width|height)=\"\d*\"\s/', "", $html );
-
-    return $html;
+function mds_get_icons_file_uri() {
+    return get_template_directory_uri() . '/assets/theme/img/icons.svg';
 }
+
+/**
+ * Get base link for load more
+ *
+ */
+function mds_get_page_link_base() {
+    $big = 999999999;
+    echo str_replace( $big.'/', '', esc_url(get_pagenum_link($big)));
+}
+
+/**
+ * Numbered Pagination
+ *
+ * @param null $the_query
+ */
+function mds_pagination($the_query = null) {
+
+    if ($the_query) {
+        $total = $the_query->max_num_pages;
+    } else {
+        global $wp_query;
+        $total = $wp_query->max_num_pages;
+    }
+
+    $big = 999999999;
+    if ($total > 1)  {
+        if(get_option('permalink_structure')) {
+            $format = 'page/%#%/';
+        } else {
+            $format = '&paged=%#%';
+        }
+        $pagination_links = paginate_links(array(
+            'base'			=> str_replace( $big, '%#%', esc_url(get_pagenum_link($big) ) ),
+            'format'		=> $format,
+            'current'		=> max( 1, get_query_var('paged') ),
+            'total' 		=> $total,
+            'mid_size'		=> 1,
+            'type' 			=> 'array',
+            'prev_text'		=> '&lt;',
+            'next_text'		=> '&gt;',
+        ));
+
+        foreach ($pagination_links as $link) {
+            echo '<li>' . $link . '</li>';
+        }
+    }
+}
+