+++ functions/woocommerce.php
@@ -12,7 +12,7 @@ remove_action( 'woocommerce_after_main_content', 'woocommerce_output_content_wra
 add_action('woocommerce_before_main_content', 'mds_woocommerce_wrapper_start', 10);
 add_action('woocommerce_after_main_content', 'mds_woocommerce_wrapper_end', 10);
 function mds_woocommerce_wrapper_start() {
-    echo '<section id="main">';
+    echo '<section id="mds-woocommerce-main">';
 }
 function mds_woocommerce_wrapper_end() {
     echo '</section>';