+++ sidebar.php
@@ -5,11 +5,11 @@
  * @package mds_starter_theme
  */
 
-if ( ! is_active_sidebar( 'sidebar-1' ) ) {
+if (!is_active_sidebar( 'sidebar-1')) {
 	return;
-}
-?>
+} ?>
 
-<div id="secondary" class="widget-area" role="complementary">
+<div class="mds-sidebar" role="complementary">
 	<?php dynamic_sidebar( 'sidebar-1' ); ?>
-</div><!-- #secondary -->
+</div>
+<!-- /.mds-sidebar -->