+++ template-parts/content-search.php
@@ -14,9 +14,7 @@
 		<?php the_title( sprintf( '<h2 class="entry-title"><a href="%s" rel="bookmark">', esc_url( get_permalink() ) ), '</a></h2>' ); ?>
 
 		<?php if ( 'post' === get_post_type() ) : ?>
-		<div class="entry-meta">
-			<?php mds_posted_on(); ?>
-		</div><!-- .entry-meta -->
+
 		<?php endif; ?>
 	</header><!-- .entry-header -->
 
@@ -25,6 +23,6 @@
 	</div><!-- .entry-summary -->
 
 	<footer class="entry-footer">
-		<?php mds_entry_footer(); ?>
+
 	</footer><!-- .entry-footer -->
 </article><!-- #post-## -->