+++ template-parts/content.php
@@ -18,9 +18,6 @@
 			}
 
 		if ( 'post' === get_post_type() ) : ?>
-		<div class="entry-meta">
-			<?php mds_posted_on(); ?>
-		</div><!-- .entry-meta -->
 		<?php
 		endif; ?>
 	</header><!-- .entry-header -->
@@ -41,6 +38,5 @@
 	</div><!-- .entry-content -->
 
 	<footer class="entry-footer">
-		<?php mds_entry_footer(); ?>
 	</footer><!-- .entry-footer -->
 </article><!-- #post-## -->